# Convert Polynomial type to struct