This PR converts the `Polynomial` from a typedef'd fixed-length array to a structure with a single member. This type is only used internally so we don't need to update any of the bindings.

PS: I will update the other types next. Wanting to do this work in small chunks.

In my opinion, a wrapped struct is safer & easier to read. See this link for details:
* https://stackoverflow.com/questions/4523497/typedef-fixed-length-array

Also, there's practically no performance difference. Benchmarked with go bindings.

Before:
```
$ ./bench.sh
goos: darwin
goarch: arm64
pkg: github.com/jtraglia/cgo-kzg-4844
BenchmarkBytesFromG1-20                         462160016               77.99 ns/op          192 B/op          2 allocs/op
BenchmarkBytesToG1-20                           89689077               401.0 ns/op           192 B/op          2 allocs/op
BenchmarkBytesToBlsField-20                     87684872               412.5 ns/op            64 B/op          2 allocs/op
BenchmarkComputeAggregateKzgProof-20                 730          49268169 ns/op             336 B/op          3 allocs/op
BenchmarkVerifyAggregateKzgProof-20                15490           2327210 ns/op             145 B/op          2 allocs/op
BenchmarkBlobToKzgCommitment-20                    18070           1996157 ns/op          131408 B/op          4 allocs/op
BenchmarkVerifyKzgProof-20                         41712            860376 ns/op             353 B/op          5 allocs/op
PASS
ok      github.com/jtraglia/cgo-kzg-4844        320.149s
```

After:
```
$ ./bench.sh
goos: darwin
goarch: arm64
pkg: github.com/jtraglia/cgo-kzg-4844
BenchmarkBytesFromG1-20                         463204014               78.04 ns/op          192 B/op          2 allocs/op
BenchmarkBytesToG1-20                           89752320               400.7 ns/op           192 B/op          2 allocs/op
BenchmarkBytesToBlsField-20                     87820915               410.6 ns/op            64 B/op          2 allocs/op
BenchmarkComputeAggregateKzgProof-20                 730          49246427 ns/op             336 B/op          3 allocs/op
BenchmarkVerifyAggregateKzgProof-20                15475           2327501 ns/op             145 B/op          2 allocs/op
BenchmarkBlobToKzgCommitment-20                    18154           1983811 ns/op          131408 B/op          4 allocs/op
BenchmarkVerifyKzgProof-20                         41888            860085 ns/op             353 B/op          5 allocs/op
PASS
ok      github.com/jtraglia/cgo-kzg-4844        320.093s
```