@jtraglia 2023-01-11
@xrchz Could we change the repo settings to squash commits for PRs?